backend/requirements.txt
fastapi
uvicorn
numpy
torch
torchaudio
openai-whisper
noisereduce
silero-vad
librosa
sentence-transformers
scikit-learn
transformers
detoxify
fasttext-wheel